+++ apps/miranum-bpmn-modeler-webview/src/types/bpmn-js.d.ts
@@ -1,22 +1,24 @@
 declare module "bpmn-js-properties-panel" {
-    export const useService: any;
+    export const useService;
 }
 
 declare module "@bpmn-io/properties-panel" {
-    export const isSelectEntryEdited: any;
-    export const SelectEntry: any;
+    export const isSelectEntryEdited;
+    export const SelectEntry;
 }
 
 declare module "camunda-bpmn-js-behaviors/lib/util/ElementUtil" {
-    export const createElement: any;
+    export const createElement;
 }
 
 declare module "@bpmn-io/element-template-chooser" {
-    const ElementTemplateChooserModule: any;
-    export default ElementTemplateChooserModule;
+    export const ElementTemplateChooserModule;
 }
 
 declare module "bpmn-js-token-simulation" {
-    const TokenSimulationModule: any;
-    export default TokenSimulationModule;
+    export const TokenSimulationModule;
+}
+
+declare module "bpmn-js-create-append-anything" {
+    export const CreateAppendElementTemplatesModule;
 }